+++ skins/oranged/templates/thumbnails_begin.php
@@ -7,6 +7,7 @@
 	<meta name="title" content="<?php $mg2->output('pagetitle');?>" />
 	<meta name="robots" content="<?php $mg2->output('robots');?>" />
 	<link href="skins/<?php $mg2->output('activeskin');?>/css/style.css" rel="stylesheet" type="text/css" />
+	<?php $mg2->ga4tag();?>
 </head>
 <body class="mg2body">
 <table cellspacing="0" cellpadding="0" class="table-top" width="100%">